==> codes/1-kzis-hierarchy.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(pdftools)
library(stringr)
```

Official hierarchy is stored in a pdf file

```{r}
hierarchy_pdf <- "../data-raw/klasyfikacja-hierarchia.pdf"
```

Read and process, group 3433 has a missing number so I add by hand

```{r}
hierachy <- pdftools::pdf_text(hierarchy_pdf)
hierachy <- lapply(hierachy, \(x) str_trim(str_split(x, "\\n")[[1]]))
hierachy <- unlist(hierachy)
hierachy[str_detect(hierachy, "Pracownicy bibliotek, galerii")] <- "3433  Pracownicy bibliotek, galerii, muzeów, informacji naukowej i pokrewni"
hierachy <- hierachy[str_detect(hierachy, "^\\d{1,6}\\s{1,}")]
hierachy <- hierachy[1:(length(hierachy)-14)]
hierachy <- str_replace(hierachy, "\\s{1,}", "\t")
hierachy <- c("class\tname", hierachy)
```

```{r}
writeLines(hierachy, con = "../data/kzis-hierarchy-2023.tsv")
```
