==> codes/4-hand-coded.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(readxl)
library(data.table)
library(stringr)
library(writexl)
```

Merging and cleaning

```{r}
hand <- dir(path = "../data-raw/hand-coded", pattern = "*done.xlsx", full.names = T)
hand <- lapply(hand, read_excel, col_types = "text")
hand <- rbindlist(hand, idcol = "ekspert", fill = T)

hand_selected <- hand[language.description == "pl", .(ekspert, 
                                                      source,
                                                      id_ogl=as.character(as.integer(ID_unit)), 
                                                      id_oferty=as.character(as.integer(id_oferty)), 
                                                      language.description, job.title, employer, region, job.type, 
                                                      position.level, category, contract.type, job.description, 
                                                      kod_zawodu = as.character(as.integer(`kod zawodu`)), 
                                                      kod_zawodu2 = as.character(as.integer(kod)),
                                                      uwagi)]

hand_selected[is.na(kod_zawodu), kod_zawodu  := kod_zawodu2 ]
hand_selected[, kod_zawodu_corrected:= kod_zawodu]
hand_selected
```

Correcting after clerical review

```{r}
hand_selected[ekspert == 1 & id_ogl == "82487"   , kod_zawodu_corrected:= "331203"]
hand_selected[ekspert == 1 & id_ogl == "186186"  , kod_zawodu_corrected:= "921490"]
hand_selected[ekspert == 1 & id_ogl == "900451"  , kod_zawodu_corrected:= "251202"]
hand_selected[ekspert == 1 & id_ogl == "190586"  , kod_zawodu_corrected:= "818990"]
hand_selected[ekspert == 1 & id_ogl == "417578"  , kod_zawodu_corrected:= "244001"]
hand_selected[ekspert == 1 & id_ogl == "831144"  , kod_zawodu_corrected:= "228203"]
hand_selected[ekspert == 1 & id_ogl == "719310"  , kod_zawodu_corrected:= "721405"]
hand_selected[ekspert == 1 & id_ogl == "972214"  , kod_zawodu_corrected:= "228203"]
hand_selected[ekspert == 1 & id_ogl == "1076160" , kod_zawodu_corrected:= "251401"]
hand_selected[ekspert == 1 & id_ogl == "630918"  , kod_zawodu_corrected:= "833202"]

hand_selected[ekspert == 2 & id_ogl == "582803" ,  kod_zawodu_corrected := "242390"]
hand_selected[ekspert == 2 & id_ogl == "1112497" , kod_zawodu_corrected := "242230"]
hand_selected[ekspert == 2 & id_ogl == "1213454" , kod_zawodu_corrected := "817290"]
hand_selected[ekspert == 2 & id_ogl == "1026821" , kod_zawodu_corrected := "311803"]
hand_selected[ekspert == 2 & id_ogl == "82487" ,   kod_zawodu_corrected := "331203"]
hand_selected[ekspert == 2 & id_ogl == "591501" ,  kod_zawodu_corrected := "121904"]
hand_selected[ekspert == 2 & id_ogl == "1166129" , kod_zawodu_corrected := "933304"]
hand_selected[ekspert == 2 & id_ogl == "792538" ,  kod_zawodu_corrected := "244001"]
hand_selected[ekspert == 2 & id_ogl == "1251772" , kod_zawodu_corrected := "333401"]
hand_selected[ekspert == 2 & id_ogl == "1237150" , kod_zawodu_corrected := "331201"]
hand_selected[ekspert == 2 & id_ogl == "940614" ,  kod_zawodu_corrected := "818990"]
hand_selected[ekspert == 2 & id_ogl == "483819" ,  kod_zawodu_corrected := "411090"]

hand_selected[ekspert == 3 & id_ogl == "186186", kod_zawodu_corrected:=  "921401"]
hand_selected[ekspert == 3 & id_ogl == "1097073", kod_zawodu_corrected:= "325512"]
hand_selected[ekspert == 3 & id_ogl == "1388500", kod_zawodu_corrected:= "911207"]
hand_selected[ekspert == 3 & id_ogl == "281182", kod_zawodu_corrected:=  "814209"]
hand_selected[ekspert == 3 & id_ogl == "977073", kod_zawodu_corrected:=  "214102"]

hand_selected[ekspert == 2	& id_ogl == "286087", kod_zawodu_corrected := "821103"]
hand_selected[ekspert == 3	& id_ogl == "628828", kod_zawodu_corrected := "321401"]

hand_selected[, id_row:=1:.N]
```


Multiple codes for one ad

```{r}
hand_selected[, possible_codes := paste(unlist(str_extract_all(uwagi, "\\d{6}")),  collapse=", "), id_row]
hand_selected[possible_codes == "NA", possible_codes:=""]
hand_selected[possible_codes !="", .(possible_codes)]
```

Correct codes to 2022 if any

```{r}
klucz <- read_excel("../data-raw/ksiz_kody_2014_2022_klucz.xls", 
                    col_names = c("zawod_old", "nazwa_old", "zawod_new", "nazwa_new"), skip = 1, col_types = "text")
klucz <- setDT(klucz)
```

```{r}
hand_selected <- merge(x = hand_selected,
                       y = klucz[zawod_old != zawod_new][, .(kod_zawodu_corrected = zawod_old,  kod_zawodu_corrected2 = zawod_new)],
                       all.x=T)

hand_selected[!is.na(kod_zawodu_corrected2), kod_zawodu_corrected:= kod_zawodu_corrected2]
hand_selected[, kod_zawodu_corrected2:=NULL]

hand_selected <- hand_selected[order(id_ogl, ekspert)]
hand_selected[, id_ogl_times :=.N, id_ogl]
hand_selected[id_ogl_times == 3, job.description := job.description[1], id_ogl]

```

Long and to train

```{r}
posible_codes_v <- hand_selected[possible_codes != "" & !is.na(kod_zawodu_corrected),  
                                 lapply(possible_codes, function(x) unlist(tstrsplit(x, ","))), 
                                   by = id_row]

hand_selected2 <- merge(x = hand_selected,
                       y = posible_codes_v,
                       by = "id_row")

hand_selected2[, kod_zawodu_corrected:= V1]
hand_selected2[, V1:= NULL]

hand_final <- rbind(hand_selected,hand_selected2)

write_xlsx(hand_final, path="../data-raw/hand-coded.xlsx")
```

```{r}
hand_final[is.na(employer) | employer == "", employer := ""]
hand_final[is.na(position.level) | position.level == "", position.level := ""]
hand_final[is.na(category) | category == "", category := ""]
hand_final[is.na(contract.type) | contract.type == "", contract.type := ""]
hand_final[is.na(job.description) | job.description == "", job.description := ""]

fwrite(
  hand_final[,.(code = kod_zawodu_corrected, desc = str_c(job.title, employer, position.level, category, job.description, sep =" "))],
  "../data/hand-coded.csv"
)
```

Processing 1k ads

```{r}
hand1k <- read_excel("../data-raw/hand-coded-1k.xlsx", 
                     col_names = c("id", "desc", "class", "name", "details", "del1", "del2"),
                     skip = 1) |> setDT()
hand1k <- hand1k[, .SD, .SDcols = id:details]
hand1k[, class:=substr(class,1,6)]

hand1k_no_add <- hand1k[is.na(details)]
hand1k_add <- hand1k[!is.na(details)]
hand1k_add[, class2:=paste0(str_extract_all(details, "\\d{6}")[[1]], collapse = ","), by = id]
hand1k_add_long <- hand1k_add[, .(class2 = unlist(tstrsplit(class2, ",", type.convert = TRUE))), by = .(id, desc, class, name, details)]

hand1k_add_long[, .(id, desc, class, class2)] |>
  melt(id.vars = c("id", "desc"), value.name = "class") |>
  {\(x) unique(x[, .(id, class, desc)])}() -> hand1k_add_long

hand1k_processed <- rbind(hand1k_no_add[,.(id, class, desc)], hand1k_add_long[, .(id, class, desc)])
hand1k_processed[class == "250102", class := "252102"]
hand1k_processed[class == "250902", class := "252902"]
hand1k_processed
```


```{r}
fwrite(x = hand1k_processed[, .(class, desc)], file = "../data/hand-coded-1k.csv", quote = T)
```
