==> codes/5-cbop.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(data.table)
library(stringr)
library(readxl)
```

```{r}
hierachy <- fread("../data/kzis-hierarchy-2023.tsv", colClasses = "character", sep = "\t")
```

```{r}
klucz <- read_excel("../data-raw/ksiz_kody_2014_2022_klucz.xls", 
                    col_names = c("zawod_old", "nazwa_old", "zawod_new", "nazwa_new"), skip = 1, col_types = "text")
klucz <- setDT(klucz)
klucz <- klucz[zawod_old != zawod_new]
```

```{r}
cbop1 <- fread("../data/cbop-train-2020-2022-no-employer.tar.gz", colClasses = "character")
cbop2 <- fread("../data/cbop-train-2022-2023-no-employer.tar.gz", colClasses = "character")
cbop <- rbind(cbop1,cbop2) |> unique()
cbop <- cbop[code != "000000"]
cbop[klucz[, .(code = zawod_old, code_new = zawod_new)], on = "code", code_new:=i.code_new]
cbop[!is.na(code_new), code := code_new]
cbop[,code_new := NULL]
cbop[, desc_nchars := nchar(desc)]

cbop[, desc_nchars_cit := cut(desc_nchars, c(0, 50, 100, 200, 300, 400, 500, Inf),
                              include.lowest = T,
                              right = T)]

cbop[,.N, keyby=desc_nchars_cit]
cbop[,.N, code]
```

Distribution is very skewed so we will sample 1 examples or 20% from each 

```{r}
set.seed(2024)
cbop_sample <- cbop[, .SD[sample(.N, max(1, round(0.2*.N)))], .(code, desc_nchars_cit)]
fwrite(x = cbop_sample[, .(code, desc)], file = "../data/cbop-for-train-test.tar.gz", quote = T)
```

```{r}
fwrite(cbop[, .N, code], "../data-raw/cbop-codes-2024.csv")
```
