==> codes/5-train-test.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(data.table)
library(stringr)
```

Official hierarchy

```{r}
hierachy <- fread("../data/kzis-hierarchy-2023.tsv", colClasses = "character", sep = "\t")
hierachy6 <- hierachy[str_detect(class, "\\d{6}")]
head(hierachy6)
```

For training only

```{r}
official <- fread("../data/kzis-occup-dictionary-long-2024.csv", colClasses = "character", 
                  col.names = c("class", "variable", "desc"))
official[nchar(class) == 5, class:=paste0("0", class)]
official[, source:="official"]
thes <- fread("../data/gus-thesaurus.csv", colClasses = "character", col.names = c("class", "desc"))
thes[, source:="gus"]
```

For both

```{r}
info <- fread("../data/kzis-occup-infodoradca-2024-updated.csv", colClasses = "character")[, source:="info"]
kprm <- fread("../data/kprm-selected.csv", colClasses = "character", col.names = c("class", "desc"))[, source:="kprm"]
hand <- fread("../data/hand-coded.csv", colClasses = "character", col.names = c("class", "desc"))[, source:="hand"]
hand_1k <- fread("../data/hand-coded-1k.csv", colClasses = "character")[, source:="hand1k"]
esco <- fread("../data/esco-kzis-selected.csv", colClasses = "character", col.names = c("class", "var", "desc"))[, source:="esco"]
cbop <- fread("../data/cbop-for-train-test.tar.gz", colClasses = "character", col.names = c("class", "desc"))[, source:="cbop"]
```

## Limited data source


Create training and test data wit 20% of CBOP data

```{r}
set.seed(123)
all_sources <- rbind(info, kprm,  hand, hand_1k, esco, cbop, fill = T)
all_sources <- all_sources[class %in% hierachy6$class] ## 8 wrong
all_sources[, size := .N, .(class, source)]
all_sources[, size_test := round(size*0.7), class]
all_sources[, id:=1:.N]

## with only one group (for testing because I will have each group in train)
test_data_ones <- all_sources[size == 1]
all_sources_rest <- all_sources[size > 1]

## sampling by size (train data)
train_data <- all_sources_rest[, .SD[sample(.N, size_test[1])], .(class, source)]
train_data <- rbind(train_data, official, thes, fill = T)

## test data bind all
test_data <- rbind(all_sources_rest[!id %in% train_data$id], test_data_ones, fill = T)

train_data <- train_data[class %in% hierachy6$class] ## 6 
test_data <- test_data[class %in% hierachy6$class]

fwrite(x = test_data[, .(class, desc, source)], file = "../data/train-test/test-data-2024.tar.gz")
fwrite(x = train_data[, .(class, desc, source)], file = "../data/train-test/train-data-2024.tar.gz")

#fwrite(x = test_data[, .(class, desc, source)], file = "../data/train-test/test-data-2024.csv", quote = T, encoding = "UTF-8")
#fwrite(x = train_data[, .(class, desc, source)], file = "../data/train-test/train-data-2024.csv", quote = T, encoding = "UTF-8")

```
