==> codes/99-coverage-of-datasets.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(data.table)
library(stringr)
```

datasets

```{r}
hierachy <- fread("../data/kzis-hierarchy-2023.tsv", colClasses = "character", sep = "\t")
hierachy6 <- hierachy[str_detect(class, "\\d{6}")]

df_full <- rbind(
  fread("../data/train-test/train-data-2024.tar.gz", colClasses = "character"),
  fread("../data/train-test/test-data-2024.tar.gz", colClasses = "character")
)
```

```{r}
hierachy6[, esco :=  class %in% unique(df_full$class[df_full$source == "esco"])]
hierachy6[, gus :=  class %in% unique(df_full$class[df_full$source == "gus"])]
hierachy6[, info :=  class %in% unique(df_full$class[df_full$source == "info"])]
hierachy6[, kprm :=  class %in% unique(df_full$class[df_full$source == "kprm"])]
hierachy6[, hand :=  class %in% unique(df_full$class[df_full$source == "hand"])]
hierachy6[, hand1k :=  class %in% unique(df_full$class[df_full$source == "hand1k"])]
hierachy6[, cbop:= class %in% unique(df_full$class[df_full$source == "cbop"])]

hierachy6[, all := (esco + gus + info + kprm + hand + hand1k + cbop) > 0]

setcolorder(hierachy6, c("class", "name", "info", "gus", "esco", "kprm", "hand", "hand1k", "cbop"))
```


By 1 digit

```{r, results='asis'}
hierachy6[, lapply(.SD, function(x) mean(x)*100), .SDcols = info:all] 
```

By 1 digit and source

```{r}
hierachy6[, lapply(.SD, function(x) mean(x)*100),  keyby=.(digit1=substr(class,1,1)), .SDcols = info:all]
```

Which are not covered


```{r}
hierachy6[all == FALSE][, .(class, name)][order(class)] |>
  fwrite("~/mac/nauka/oja-lab/job-ads-datasets/data/codes-not-coveted.csv")
```
